+++ Autodesk.PackageBuilder.Tests/Data/DataBuild_Tests.cs
@@ -1,6 +1,7 @@
 ﻿using Autodesk.PackageBuilder.Model.Application;
 using Autodesk.PackageBuilder.Tests.Utils;
 using NUnit.Framework;
+using System.Xml.Serialization;
 
 namespace Autodesk.PackageBuilder.Tests.Data
 {
@@ -57,5 +58,51 @@ namespace Autodesk.PackageBuilder.Tests.Data
                 builder.AssertElementAttribute("ComponentEntry", "Attribute", i);
             }
         }
+
+        [Test]
+        public void Build_CustomExtensibleData()
+        {
+            var builder = BuilderUtils.Build<PackageContentsBuilder>(builder =>
+            {
+                var custom = new CustomElement()
+                {
+                    Name = "Name",
+                    Value = "Value"
+                };
+
+                var componentRevit2021 = builder.Components
+                    .CreateEntry("Revit 2021");
+
+                componentRevit2021.DataBuilder.CreateAttribute("Attribute", true);
+                componentRevit2021.DataBuilder.CreateElement("Element", true);
+                componentRevit2021.DataBuilder.CreateElement("CustomElement", custom);
+            });
+
+            var result = builder.ToString();
+            System.Console.WriteLine(result);
+
+            Assert.AreEqual(Expected, result, $"Expected: {Expected}\nContent: {result}");
+        }
+
+        public class CustomElement : ExtensibleData
+        {
+            [XmlAttribute]
+            public string Name { get; set; }
+            [XmlElement]
+            public string Value { get; set; }
+        }
+
+        public static string Expected => """"
+            <?xml version="1.0" encoding="utf-8"?>
+            <ApplicationPackage>
+              <CompanyDetails />
+              <Components Description="Revit 2021" Attribute="True">
+                <Element>true</Element>
+                <CustomElement Name="Name">
+                  <Value>Value</Value>
+                </CustomElement>
+              </Components>
+            </ApplicationPackage>
+            """";
     }
 }