+++ Autodesk.PackageBuilder/Builder/Abstractions/DataBuilderBase.cs
@@ -1,26 +1,26 @@
 ﻿namespace Autodesk.PackageBuilder
 {
     /// <summary>
-    /// Provides a base builder for constructing and manipulating <see cref="DataBase"/> objects.
+    /// Provides a base builder for constructing and manipulating <see cref="ExtensibleData"/> objects.
     /// </summary>
     /// <remarks>
     /// This builder offers a fluent API for creating and modifying attributes and elements
-    /// within a <see cref="DataBase"/> instance. It supports both direct manipulation of the
+    /// within a <see cref="ExtensibleData"/> instance. It supports both direct manipulation of the
     /// underlying data model and creation of nested elements or attributes using generic overloads.
     /// </remarks>
-    public class DataBuilderBase : SingleBuilderBase<DataBuilderBase, DataBase>
+    public class DataBuilderBase : SingleBuilderBase<DataBuilderBase, ExtensibleData>
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="DataBuilderBase"/> class with the specified <see cref="DataBase"/> model.
+        /// Initializes a new instance of the <see cref="DataBuilderBase"/> class with the specified <see cref="ExtensibleData"/> model.
         /// </summary>
-        /// <param name="model">The <see cref="DataBase"/> instance to wrap and build upon.</param>
-        public DataBuilderBase(DataBase model)
+        /// <param name="model">The <see cref="ExtensibleData"/> instance to wrap and build upon.</param>
+        public DataBuilderBase(ExtensibleData model)
         {
             SetDataInternal(model);
         }
 
         /// <summary>
-        /// Creates a new data attribute with the specified name and value in the underlying <see cref="DataBase"/>.
+        /// Creates a new data attribute with the specified name and value in the underlying <see cref="ExtensibleData"/>.
         /// </summary>
         /// <param name="name">The name of the attribute to create.</param>
         /// <param name="value">The value of the attribute.</param>
@@ -42,33 +42,33 @@
         }
 
         /// <summary>
-        /// Creates a new data element with the specified name and value in the underlying <see cref="DataBase"/>.
+        /// Creates a new data element with the specified name and value in the underlying <see cref="ExtensibleData"/>.
         /// </summary>
         /// <param name="name">The name of the data element to create.</param>
         /// <param name="value">
-        /// The value to associate with the new data element. If the value is a <see cref="DataBase"/>, it will be used to create a nested element.
+        /// The value to associate with the new data element. If the value is a <see cref="ExtensibleData"/>, it will be used to create a nested element.
         /// Otherwise, the value will be set directly.
         /// </param>
         /// <returns>
-        /// A new <see cref="DataBuilderBase"/> instance for the created data element if <paramref name="value"/> is a <see cref="DataBase"/>; otherwise, <c>null</c>.
+        /// A new <see cref="DataBuilderBase"/> instance for the created data element if <paramref name="value"/> is a <see cref="ExtensibleData"/>; otherwise, <c>null</c>.
         /// </returns>
         public DataBuilderBase CreateElement(string name, object value)
         {
-            if (value is DataBase dataBase)
-                return new DataBuilderBase(Data.CreateDataElement(name, dataBase));
+            if (value is ExtensibleData extensibleData)
+                return new DataBuilderBase(Data.CreateDataElement(name, extensibleData));
 
             Data.CreateDataElement(name, value);
             return null;
         }
 
         /// <summary>
-        /// Creates a new data attribute with the specified name and value in a new instance of <typeparamref name="T"/> derived from <see cref="DataBase"/>.
+        /// Creates a new data attribute with the specified name and value in a new instance of <typeparamref name="T"/> derived from <see cref="ExtensibleData"/>.
         /// </summary>
-        /// <typeparam name="T">The type of <see cref="DataBase"/> to use for creating the attribute. Must have a parameterless constructor.</typeparam>
+        /// <typeparam name="T">The type of <see cref="ExtensibleData"/> to use for creating the attribute. Must have a parameterless constructor.</typeparam>
         /// <param name="name">The name of the attribute to create.</param>
         /// <param name="value">The value of the attribute.</param>
         /// <returns>The current <see cref="DataBuilderBase"/> instance for method chaining.</returns>
-        public DataBuilderBase CreateAttribute<T>(string name, object value) where T : DataBase, new()
+        public DataBuilderBase CreateAttribute<T>(string name, object value) where T : ExtensibleData, new()
         {
             var data = GetNewPropertyValue<T>(Data);
             data.CreateDataAttribute(name, value);
@@ -76,32 +76,32 @@
         }
 
         /// <summary>
-        /// Creates a new data element with the specified name in a new instance of <typeparamref name="T"/> derived from <see cref="DataBase"/>.
+        /// Creates a new data element with the specified name in a new instance of <typeparamref name="T"/> derived from <see cref="ExtensibleData"/>.
         /// </summary>
-        /// <typeparam name="T">The type of <see cref="DataBase"/> to use for creating the element. Must have a parameterless constructor.</typeparam>
+        /// <typeparam name="T">The type of <see cref="ExtensibleData"/> to use for creating the element. Must have a parameterless constructor.</typeparam>
         /// <param name="name">The name of the data element to create.</param>
         /// <returns>A new <see cref="DataBuilderBase"/> instance for the created data element.</returns>
-        public DataBuilderBase CreateElement<T>(string name) where T : DataBase, new()
+        public DataBuilderBase CreateElement<T>(string name) where T : ExtensibleData, new()
         {
             var data = GetNewPropertyValue<T>(Data);
             return new DataBuilderBase(data.CreateDataElement(name));
         }
 
         /// <summary>
-        /// Creates a new data element with the specified name and value in a new instance of <typeparamref name="T"/> derived from <see cref="DataBase"/>.
+        /// Creates a new data element with the specified name and value in a new instance of <typeparamref name="T"/> derived from <see cref="ExtensibleData"/>.
         /// </summary>
-        /// <typeparam name="T">The type of <see cref="DataBase"/> to use for creating the element. Must have a parameterless constructor.</typeparam>
+        /// <typeparam name="T">The type of <see cref="ExtensibleData"/> to use for creating the element. Must have a parameterless constructor.</typeparam>
         /// <param name="name">The name of the data element to create.</param>
-        /// <param name="value">The value to associate with the new data element. If the value is a <see cref="DataBase"/>, it will be used to create a nested element.</param>
+        /// <param name="value">The value to associate with the new data element. If the value is a <see cref="ExtensibleData"/>, it will be used to create a nested element.</param>
         /// <returns>
-        /// A new <see cref="DataBuilderBase"/> instance for the created data element if <paramref name="value"/> is a <see cref="DataBase"/>; otherwise, <c>null</c>.
+        /// A new <see cref="DataBuilderBase"/> instance for the created data element if <paramref name="value"/> is a <see cref="ExtensibleData"/>; otherwise, <c>null</c>.
         /// </returns>
-        public DataBuilderBase CreateElement<T>(string name, object value) where T : DataBase, new()
+        public DataBuilderBase CreateElement<T>(string name, object value) where T : ExtensibleData, new()
         {
             var data = GetNewPropertyValue<T>(Data);
 
-            if (value is DataBase dataBase)
-                return new DataBuilderBase(data.CreateDataElement(name, dataBase));
+            if (value is ExtensibleData extensibleData)
+                return new DataBuilderBase(data.CreateDataElement(name, extensibleData));
 
             data.CreateDataElement(name, value);
             return null;